Implement hotel reservation management and itinerary integration
- Add DELETE endpoint to `HotelAPIService` for canceling reservations
- Update `Trip` model and database to support hotel reservation data
- Implement `addTripFromBooking` in `TripViewModel` to automatically create trips from hotel bookings
- Prevent editing or deleting itinerary items that originate from a reservation
- Add a "My Reservations" screen and update navigation logic
- Sync Firebase user email with local database during login and registration
- Update Room database version and improve conflict strategies in `UserDao`

+++ app/app/src/main/java/com/monarca/smarttravel/domain/interfaces/TripRepository.kt
@@ -10,7 +10,7 @@ interface TripRepository {
     fun getAllTrips(): Flow<List<Trip>>
     fun getTripsByUser(userId: String): Flow<List<Trip>>
     suspend fun getTripById(tripId: Int): Trip?
-    suspend fun addTrip(trip: Trip): Int
+    suspend fun addTrip(trip: Trip): Long
     suspend fun updateTrip(trip: Trip): Int
     suspend fun deleteTrip(tripId: Int): Int
     suspend fun updateImage(tripId: Int, newImageResId: String?): Int
